+++ src/executeVirtualCommand/buildFile.ts
@@ -4,25 +4,27 @@ import { asOutputFile, buildInputFile } from '../util'
 import { VirtualCommandContext, VirtualCommand } from './VirtualCommand'
 
 export default {
-  name: 'buildInputFile',
+  name: 'buildFile',
   predicate(c: VirtualCommandContext): boolean {
-    return c.command[0] === 'buildInputFile'
+    return c.command[0] === 'buildFile'
   },
   async execute(c: VirtualCommandContext): Promise<ExecuteResult> {
-    const outputFiles: MagickOutputFile[] = []
+    let outputFile: MagickOutputFile
     const stderr = []
     try {
-      outputFiles.push(await asOutputFile(await buildInputFile(c.command[1], c.command[2] || undefined)))
+      outputFile = {...await asOutputFile(await buildInputFile(c.command[1], c.command[2] || undefined)), ignore: true}
     } catch (error) {
       stderr.push(error.errorMessage || error + '')
     }
     const result: ExecuteResult = {
-      outputFiles,
+      outputFiles: outputFile ? [outputFile]: [],
       commands: [c.command],
       command: c.command,
       exitCode: 0,
-      stderr, stdout: outputFiles.length ? [unquote(outputFiles[0].name)] : [],
-      inputFiles: values(c.files), results: [],
+      stderr,
+      stdout: outputFile ? [unquote(outputFile.name)] : [],
+      inputFiles: values(c.files),
+      results: [],
     }
     return { ...result, results: [result] }
   },